+++ toucan_connectors/http_api/http_api_connector.py
@@ -20,6 +20,7 @@ try:
         NoopPaginationConfig,
         extract_pagination_info_from_result,
     )
+    from toucan_connectors.utils.dataframe import append_dataframes
 
     CONNECTOR_OK = True
 except ImportError as exc:  # pragma: no cover
@@ -194,7 +195,7 @@ class HttpAPIConnector(ToucanConnector, data_source_model=HttpAPIDataSource):
                 raise
         if data_source.flatten_column:
             dfs = [json_to_table(df, columns=[data_source.flatten_column]) for df in dfs]
-        return functools.reduce(lambda df, df_second: df.append(df_second), dfs)
+        return functools.reduce(append_dataframes, dfs)
 
     def _render_query(self, data_source):
         query = nosql_apply_parameters_to_query(